♻️  extract commonStyles
