+++ src/containers/NewContractMember/NewContractMemberSupplyPoint.jsx
@@ -11,6 +11,7 @@ import CUPS from '../../components/CUPS'
 import AlertBox from '../../components/AlertBox'
 
 import { iconRequirements } from '../../themes/commonStyles'
+import { iconOffRequirements } from '../../themes/commonStyles'
 
 import Grid from '@mui/material/Grid'
